# Optimize `BuildPhaseGenerator`'s method of detecting assets and localized files